Refactor to shorten ContainerModelBuilderTest
* Move jvm args tests into new subclass
* Move test logger to ContainerModelBuilderTestBase for reuse
